requirements.txt
appdirs==1.4.4
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cssselect==1.1.0
fake-useragent==0.1.11
Flask==2.0.2
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
parse==1.19.0
pyee==8.2.2
pyppeteer==0.2.6
pyquery==1.4.3
python-dotenv==0.19.1
requests==2.26.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.2.1
tqdm==4.62.3
urllib3==1.26.7
w3lib==1.22.0
websockets==9.1
Werkzeug==2.0.2
zipp==3.6.0
